Put second-level class under the class they are used; Achieved network connection
